\name{LIHNPSD_plot_std4gr}
\alias{LIHNPSD_plot_std4gr}
\title{
  Standard utility to plot a 4-chart graph on a given data set and PSD fit
}
\description{
  Standard utility to plot a 4-chart graph on a given data set and PSD fit.
}
\usage{
LIHNPSD_plot_std4gr(th, dt, EPS, file = NA)
}
\arguments{
  \item{th}{ The theoretical object from \code{LIHNPSD_theoretical_result} }
  \item{dt}{ The data set object from \code{LIHNPSD_prepare_data} }
  \item{EPS}{ TRUE: plot for eps file; FALSE: plot to screen }
  \item{file}{ File name of the eps output }
}
\value{ No output value }
\keyword{ plot }
