\name{LIHNPSD_UnitTest}
\alias{LIHNPSD_UnitTest}
\title{
  Perform major unit tests on the package 
}
\description{
  Perform major unit tests on the package 
}
\usage{
LIHNPSD_UnitTest(mpfr = 0)
}
\arguments{
  \item{mpfr}{ If non-zero value is specified, MPFR will be used. }
}
\value{ Error if not passed }

\seealso{
\code{\link{psdunittest}}
}
\keyword{ testing }
