\name{poisson_sum}
\alias{poisson_sum}
\title{
  Utility to calculate the Poisson sum
}
\description{
  Utility to calculate the Poisson sum
}
\usage{
poisson_sum(d, fn)
}
\arguments{
  \item{d}{ A fully specified PSD S3 object }
  \item{fn}{ The input function }
}
\value{ The numeric value of the sum on \code{fn} }

\keyword{ Poisson sum }
