\name{psdkernel}
\alias{psdkernel}
\title{
  Calculate the ratio of the actual volatility vs the unit volatility
}
\description{
  Calculate the ratio of the actual volatility vs the unit volatility
}
\usage{
psdkernel(d, k)
}
\arguments{
  \item{d}{ A fully specified PSD S3 object }
  \item{k}{ The k in the Poisson distribution }
}
\value{ The numeric value of the kernel }
