\name{psdkurtosis}
\alias{psdkurtosis}
\title{
  Calculate the kurtosis based on Poisson sum of moments
}
\description{
  Calculate the kurtosis based on Poisson sum of moments.
}
\usage{
psdkurtosis(d)
}
\arguments{
  \item{d}{ A fully specified PSD S3 object }
}
\value{ Numeric value of the kurtosis }
\references{
See Section "The Kurtosis" in the PSD paper
}

\seealso{
\code{\link{rawmu4}}
}
\keyword{ statistics }
