\name{psdskewness}
\alias{psdskewness}
\title{
  Calculate the skewness based on Poisson sum of moments
}
\description{
  Calculate the skewness based on Poisson sum of moments.
}
\usage{
psdskewness(d)
}
\arguments{
  \item{d}{ A fully specified PSD S3 object }
}
\value{ Numeric value of the skewness }
\references{
See Section "The Skewness" in the PSD paper
}

\seealso{
\code{\link{rawmu3}}
}
\keyword{ statistics }
