\name{psdunittest}
\alias{psdunittest}
\title{
  Perform unit tests on the specified PSD S3 object
}
\description{
  Perform unit tests on the specified PSD S3 object
}
\usage{
psdunittest(d)
}
\arguments{
  \item{d}{ A PSD S3 object }
}
\value{ Error if not passed }

\seealso{
\code{\link{LIHNPSD_UnitTest}}
}
\keyword{ testing }
