\name{psdvariance}
\alias{psdvariance}
\title{
  Calculate the variance based on Poisson sum of moments
}
\description{
  Calculate the variance based on Poisson sum of moments.
}
\usage{
psdvariance(d)
}
\arguments{
  \item{d}{ A fully specified PSD S3 object }
}
\value{ Numeric value of the variance }
\references{
See Section "The Variance" in the PSD paper
}

\seealso{
\code{\link{rawmu2}}
}
\keyword{ statistics }
