\name{rawcdf}
\alias{rawcdf}
\title{
  Calculate the raw CDF
}
\description{
  Calculate the raw CDF
}
\usage{
rawcdf(d, x)
}
\arguments{
  \item{d}{ A fully specified PSD S3 object }
  \item{x}{ x of CDF }
}
\value{Return CDF(x)}
\references{
See Section "Development of PSD" in the PSD paper
}
\seealso{
\code{\link{rawcdfinv}}
}
\keyword{ CDF }
