\name{rawdensity0}
\alias{rawdensity0}
\title{
  The raw PDF at x=0
}
\description{
  The raw PDF at x=0. It is implemented from the analytic result 
  primarily for internal validation purpose.
}
\usage{
rawdensity0(d)
}
\arguments{
  \item{d}{ A fully specified PSD S3 object }
}
\value{Return PDF(x)}
\references{
See Section "Development of PSD" in the PSD paper
}
\seealso{
\code{\link{rawdensity}}
}
\keyword{ PDF }
