\name{rawdensityslope}
\alias{rawdensityslope}
\title{
  Calculate the slope of the PDF (dP/dx)
}
\description{
  Calculate the slope of the PDF (dP/dx)
}
\usage{
rawdensityslope(d, x)
}
\arguments{
  \item{d}{ A fully specified PSD S3 object }
  \item{x}{ x of dP/dx }
}
\value{Return dP/dx}
\references{
See Section "Tail Index" in the PSD paper
}
\seealso{
\code{\link{rawdensity}}
}
\keyword{ PDF }
