\name{rawmu1}
\alias{rawmu1}
\title{
  Calculate the Poisson sum of the first moment (mean) of PSD without location parameter
}
\description{
  Calculate the Poisson sum of the first moment (mean) of PSD without location parameter.
}
\usage{
rawmu1(d)
}
\arguments{
  \item{d}{ A fully specified PSD S3 object }
}
\value{ Numeric value of the mean }
\references{
See Section "The Mean" in the PSD paper
}

\seealso{
\code{\link{rawmean}}, \code{\link{mu1_analytic}}
}
\keyword{ statistics }
