\name{rawmu3}
\alias{rawmu3}
\title{
  Calculate the Poisson sum of the third moment of PSD without location parameter
}
\description{
  Calculate the Poisson sum of the third moment of PSD without location parameter.
}
\usage{
rawmu3(d)
}
\arguments{
  \item{d}{ A fully specified PSD S3 object }
}
\value{ Numeric value of the third moment }
\references{
See Section "The Skewness" in the PSD paper
}

\seealso{
\code{\link{psdskewness}}, \code{\link{mu3_analytic}}
}
\keyword{ statistics }
