\name{tailindex}
\alias{tailindex}
\title{
  Calculate the tail index
}
\description{
  Calculate the tail index
}
\usage{
tailindex(d, x)
}
\arguments{
  \item{d}{ A fully specified PSD S3 object }
  \item{x}{ x where the tail index is evaluated }
}
\value{Return the tail index}
\references{
See Section "Tail Index" in the PSD paper
}
\keyword{ tail }
