\name{tailindex_plot}
\alias{tailindex_plot}
\title{
  Generate the plot of tail index for the specified PSD
}
\description{
  Generate the plot of tail index for the specified PSD
}
\usage{
tailindex_plot(d, xmin, xmax, ymax=0.4)
}
\arguments{
  \item{d}{ A fully specified PSD S3 object }
  \item{xmin}{ Minimum of x-axis }
  \item{xmax}{ Maximum of x-axis }
  \item{ymax}{ Maximum of y-axis, default is 0.4 }
}
\value{ N/A }
\references{
See Section "Tail Index" in the PSD paper
}
\seealso{
\code{\link{tailindex}}
}
\keyword{ tail }
